+++ main/src/main/kotlin/io/void/html/Element.kt
@@ -1,4 +1,4 @@
-package io.void.html.element
+package io.void.html
 
 import io.void.html.attributes.Attribute
 import io.void.html.attributes.AttributeNames
